feat: seller register form
